+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 * Core: add `requestVoid` method to network client.
 * Behaviours: add keyboard avoiding behaviour (using the system provided layout guide).
 * XcodeGen: added a template for shared frameworks.
+* Added support for `Combine` in both `Core` & `CoreData`.
 
 ## [2.0.3](https://github.com/appwise-labs/AppwiseCore/releases/tag/2.0.3)